# Define attributes for channel and user channel model