The impact is rather minimal but still something, I guess.

Before:
```
Run 1:   624 requests in 10.11s, 74.99MB read
Run 2:   711 requests in 10.14s, 85.45MB read
Run 3:   685 requests in 10.09s, 82.32MB read
Run 4:   687 requests in 10.10s, 82.56MB read
Run 5:   680 requests in 10.08s, 81.72MB read
Run 6:   696 requests in 10.06s, 83.65MB read
Run 7:   669 requests in 10.09s, 80.40MB read
Run 8:   706 requests in 10.06s, 84.85MB read
Run 9:   696 requests in 10.14s, 83.65MB read
Run 10:   682 requests in 10.14s, 81.98MB read
Final run:
Running 10s test @ https://on.nanaya.pro/community/chat
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   166.66ms   25.88ms 545.51ms   96.67%
    Req/Sec    18.00      7.29    30.00     49.44%
  717 requests in 10.06s, 86.17MB read
Requests/sec:     71.27
Transfer/sec:      8.57MB
```

After:
```
Run 1:   723 requests in 10.08s, 86.89MB read
Run 2:   755 requests in 10.06s, 90.74MB read
Run 3:   774 requests in 10.08s, 93.02MB read
Run 4:   777 requests in 10.08s, 93.38MB read
Run 5:   753 requests in 10.17s, 90.50MB read
Run 6:   759 requests in 10.06s, 91.22MB read
Run 7:   747 requests in 10.07s, 89.78MB read
Run 8:   734 requests in 10.08s, 88.21MB read
Run 9:   757 requests in 10.07s, 90.98MB read
Run 10:   754 requests in 10.05s, 90.62MB read
Final run:
Running 10s test @ https://on.nanaya.pro/community/chat
  4 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   160.23ms   34.53ms 638.27ms   95.63%
    Req/Sec    18.66      7.61    30.00     43.60%
  751 requests in 10.06s, 90.26MB read
Requests/sec:     74.66
Transfer/sec:      8.97MB
```